* Fix wrong type used for soundfx priority. From the comment in fxset.h it is clear that the type should be an integer with range 0 .. 255, which means uint8_t. Especially the type should not be signed, which it was in ActPlayFX (production program command). This bug caused a compilation warning about comparing signed and unsigned expressions. The bug originates from the use of ProductionAction::iparam1 for storing soundfx priority in revisions prior to 3373. * Fix wrong comment in Carrier::find_pending_item. * Remove some excessive debug output.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3458 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/production_program.h
@@ -362,7 +362,7 @@ struct ActPlayFX : public ProductionAction {
 	virtual void execute(Game &, ProductionSite &) const;
 private:
 	std::string name;
-	int32_t     priority;
+	uint8_t     priority;
 };
